+++ frontend/src/pages/create-workflow/sidebar.tsx
@@ -1,4 +1,4 @@
-import { useDnD } from "@/hooks/dnd-context";
+import { useDnD } from "@/pages/create-workflow/context/dnd-context";
 import { DragEvent } from "react";
 import { Box, Divider, Typography } from "@mui/material";
 import PlayCircleIcon from "@mui/icons-material/PlayCircle";
